backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.18
httpx==0.28.1
docker==7.1.0
minio==7.2.12
psutil==6.1.1
pydantic==2.10.4
